EGE2022_05: Move N05.py to EGE2022 folder
