+++ EGE/Gen/EGE2022/N05.py
